+++ memdocs/wizards/__init__.py
@@ -1,0 +1,5 @@
+"""Automated wizards for Empathy-DocInt integration."""
+
+from .empathy_sync import EmpathySyncWizard
+
+__all__ = ["EmpathySyncWizard"]